Support Traject 3.0.0
Note that this is a rather naïve commit to 1) use Traject 3.0.0, and 2) make tests pass. We might want to do further work on this branch to more fully support the changes in Traject 3, and eliminate any code in TrajectPlus that is duplicate of what is it in Traject 3. This commit makes no attempt to maintain backwards compatibility with past versions of TrajectPlus.

This is mostly for discussion purposes.